feat(web): implement logic for project invitation
