+++ apps/web/src/features/project/model/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useInviteProject } from './useInviteProject';
