+++ apps/web/src/features/project/model/index.ts
@@ -1,1 +1,2 @@
 export * from './constants';
+export * from './hooks';
