+++ apps/web/src/shared/api/request.ts
@@ -2,6 +2,7 @@ export const REQUEST = {
   LOGIN: '/login',
   SIGNUP: '/v1/user/join',
   PROJECT: '/v1/project',
+  INVITE_PROJECT: '/v1/project/invite',
   REFRESH: '/v1/user/refresh',
   CHECK_VALID_ID: '/v1/user/valid/id',
   CHECK_VALID_EMAIL_CODE: '/v1/email/auth',
